Lablist fix, decimal break fix for MP

+++ DiceBot/moneypot.cs
@@ -352,7 +352,7 @@ namespace DiceBot
         public override decimal GetLucky(string server, string client, int nonce)
         {
             long cl = long.Parse(client);
-            double roll = Math.Floor((100.0 / 4294967296.0) * (double)((long)(((double)nonce) + cl) % (long)(4294967296)) * (long)100) / (long)100;
+            decimal roll = Math.Floor((100.0m / 4294967296.0m) * (decimal)((long)(((decimal)nonce) + cl) % (long)(4294967296)) * (long)100) / (long)100;
             return roll;
         }
     }
